Moved xrfi performance tests to scripts
